Fix a spacing issue and two bugs. Add milestones.
By spacing, I mean there was an extra space where there shouldn't have been
and it was causing a formatting error. Also, in
@sigmavirus24/Todo.txt-python#9 I mentioned that there should be
a #{milestone - date} part to the issue item. So I added that in this
commit.

The first bug was not setting url when formatting it. The second was that I
was missing a terminating parenthesis.