Minor changes to IssuesParser. Change caching.
IssuesParser.print_issues() was becoming more of a formatting function so I
changed it to reflect that and made a separate print_issues() function for
it. It now stores the formatted strings in a list.

Caching was done in a sort of lazy way. It now checks to make sure it is
less than an hour old before using the file. If it's an hour old or more,
it'll go to GitHub to get the issues.

I'm debating whether to add a check to see if the issue is also a pull
request and how to denote that. I don't like my original idea of using a
label, i.e. '# title (submitter) +label1 +label2 +Pull #{milestone - date}'

I'm thinking something like the following might be better:
'#/p title (submitter) +label1 #{milestone - date}'
'#/i title (submitter) +label1 #{milestone - date}'
Where the /p would denote a pull request and the /i an issue.

Also, I'm debating whether or not I should replace spaces in project labels
with underscores, i.e. in kennethreitz/clint:
'  2 clint.textui.maxwidth context manager (kennethreitz) +Feature Request'
would become
'  2 clint.textui.maxwidth context manager (kennethreitz) +Feature_Request'