A bit of refactoring.
I realized that (in Python 3, and unfortunately not in Python 2) the return
value of a urlopen call is an instance of 'HTTPResponse'. So I thought I could
make GitHub a descendant of that class. Alas, Python 2 has no such class and the
return value is an instance of addinfourl instead and doesn't have all of the
methods associated with HTTPResponse. Were this entirely for Python 3, the code
would be incredibly nicer because of a lot of the functions I could inherit.
Alas, keeping backwards compatibility is suddenly sub-optimal in my opinion.

This has forced me to refactor a bit of the code. Not much. Some of it is just
more intelligent. I still need to majorly refactor the classes because I think
both GitHub and Issues are keeping track of the user/password data. Issues
doesn't need to do this though.