Started the hard work.
Set up a syntax for the configuration file. We need the option to have a value
and for now, I'll ignore it. I reserve the right to start to use it in the
future.

I will be writing the list_all() function next because I expect that to be the
most popular. It will be the default 'action'. I'll probably break down the
majority of that work into other functions so that they can be reused
elsewhere.